test_upstream_compatibility: handle pkgrel offset
A few pmaports are generated from Alpine's aports, using the same pkgver
and pkgrel. With pmbootstrap!1806, we introduce a pkgrel offset, that
allows increasing the pkgrel independently of changes in Alpine's
pkgrel. The original pkgrel is stored in _pkgrel, so adjust the upstream
compatibility test to read _pkgrel for comparing, if it is present.

Depends: pmbootstrap!1806